feat: run MCP and REST API servers in background when TUI starts
When `fungible` launches, it now automatically starts an HTTP MCP server
(StreamableHTTP, port 3741) and the REST API (port 3456) in the same
process alongside the TUI. All three share a single db connection so
MCP/API writes are immediately visible on the next TUI navigation.

Also adds `fungible mcp` and `fungible api` subcommands to bin/fungible
for running either server standalone without the TUI.

Internals: extracted createMcpServer() factory into mcp/create-server.ts
so the tool registrations are shared between the stdio and HTTP transports.
Refactored api/server.ts to export startApiServer() with an isMain guard
for standalone use.
