+++ tui/index.tsx
@@ -10,6 +10,8 @@ import { syncAll } from '../core/sync.js';
 import { rebuildDisplayNames } from '../core/rename.js';
 import { App } from './App.js';
 import { Setup } from './Setup.js';
+import { startMcpHttpServer } from '../mcp/http.js';
+import { startApiServer } from '../api/server.js';
 
 const isDemo = process.argv.includes('--demo');
 
@@ -25,5 +27,11 @@ if (process.argv.includes('--setup')) {
   }
   await rebuildDisplayNames();
   if (!isDemo) syncAll().catch(() => {});
+
+  const mcpPort = parseInt(process.env.FUNGIBLE_MCP_PORT ?? '3741', 10);
+  const apiPort = parseInt(process.env.FUNGIBLE_API_PORT ?? '3456', 10);
+  startMcpHttpServer(mcpPort);
+  startApiServer(apiPort, { quiet: true });
+
   render(<App />);
 }
